11 - For Post creation, update, deletion, the userId is the ID of the current user

+++ Agora.API/Controllers/PostsController.cs
@@ -1,4 +1,5 @@
-﻿using Agora.API.DTOs.Post;
+﻿using System.Security.Claims;
+using Agora.API.DTOs.Post;
 using Agora.API.InputValidation.Interfaces;
 using Agora.API.Orchestrators.Interfaces;
 using Agora.API.QueryParams;
@@ -6,6 +7,7 @@ using Agora.Core.Enums;
 using Agora.Core.Interfaces;
 using Agora.Core.Models;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Agora.API.Controllers;
@@ -20,6 +22,7 @@ public class PostsController(
     : ControllerBase
 {
     private const string PostNotFoundMessage = "Post not found.";
+    private const string NotOwnerMessage = "Current user is not the owner of the post.";
 
     [HttpGet]
     public async Task<ActionResult<IReadOnlyList<PostSummaryDto>>> GetAllPosts([FromQuery] PostQueryParameters queryParameters)
@@ -38,6 +41,7 @@ public class PostsController(
             : Ok(mapper.Map<PostDetailsDto>(post));
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<PostDetailsDto>> CreatePost([FromBody] CreatePostDto postDto)
     {
@@ -51,10 +55,18 @@ public class PostsController(
         {
             return BadRequest(new { Errors = inputErrors });
         }
+        
+        // Assign userId of current user to the post
+        string? currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (currentUserId != null)
+        {
+            return Unauthorized("User ID not found in claims.");
+        }
 
         // Transform to the full entity and validate with business rules
         Post post = mapper.Map<Post>(postDto);
         post.Status = PostStatus.Draft;
+        post.UserId = currentUserId!;
         
         IList<string> businessRulesErrors = await businessRulesValidationOrchestrator.ValidateAndProcessPostAsync(post);
         if (businessRulesErrors.Count != 0)
@@ -82,6 +94,7 @@ public class PostsController(
         return BadRequest("Problem creating the post.");
     }
 
+    [Authorize]
     [HttpPut("{id:long}")]
     public async Task<ActionResult> UpdatePost([FromRoute] long id, [FromBody] UpdatePostDto postDto)
     {
@@ -96,13 +109,20 @@ public class PostsController(
             return NotFound(PostNotFoundMessage);
         }
         
+        // Ownership validation
+        string? currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (existingPost.UserId != currentUserId)
+        {
+            return Unauthorized(NotOwnerMessage);
+        }
+        
         // Input validation
         List<string> inputErrors = await inputValidator.ValidateInputPostDtoAsync(postDto);
         if (inputErrors.Count != 0)
         {
             return BadRequest(new { Errors = inputErrors });
         }
-
+        
         // Transform to the full entity and validate with business rules
         Post post = mapper.Map<Post>(postDto);
         IList<string> businessRulesErrors = await businessRulesValidationOrchestrator.ValidateAndProcessPostAsync(post);
@@ -119,6 +139,7 @@ public class PostsController(
             : BadRequest("Problem updating the post.");
     }
 
+    [Authorize]
     [HttpDelete("{id:long}")]
     public async Task<ActionResult> DeletePost([FromRoute] long id)
     {
@@ -127,6 +148,13 @@ public class PostsController(
         if (post == null)
         {
             return NotFound(PostNotFoundMessage);
+        }
+        
+        // Ownership validation
+        string? currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (post.UserId != currentUserId)
+        {
+            return Unauthorized(NotOwnerMessage);
         }
 
         repo.DeletePost(post);
